refactor: Add browse_tests tests to test_browse_scrape_text tests
